Intersting_code_assistant/requirements.txt
streamlit==1.28.0
python-dotenv==1.0.0
requests==2.31.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy>=1.24.3
torch==2.5.0